blog/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.15.1
mysql-connector-python==8.0.15
bcrypt==3.1.6
Flask-WTF==0.14.2